+++ server/types/app.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	dbm "github.com/cosmos/cosmos-db"
+
 	"github.com/cosmos/gogoproto/grpc"
 	"github.com/spf13/cobra"
 	abci "github.com/tendermint/tendermint/abci/types"
@@ -16,7 +17,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/server/api"
 	"github.com/cosmos/cosmos-sdk/server/config"
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 )
 
 // ServerStartTime defines the time duration that the server need to stay running after startup
@@ -58,7 +59,7 @@ type (
 		RegisterNodeService(client.Context)
 
 		// CommitMultiStore return the multistore instance
-		CommitMultiStore() sdk.CommitMultiStore
+		CommitMultiStore() storetypes.CommitMultiStore
 	}
 
 	// AppCreator is a function that allows us to lazily initialize an